requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fake-useragent==1.5.1
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
numpy==2.1.3
pandas==2.2.3
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
yarl==1.17.1
